Allow compiler optimisations with -fno-delete-null-pointer-checks
If at least -O1 is given then the client crashes on startup without
disabling this optimisation. It is supported at least since GCC 3.4.

This was a tricky one to find because -O0 works and -O1 doesn't but
the GCC manual says this optimisation is only enabled at -O2. Turns
out it is also enabled by default on most targets.

It may be possible to narrow it down to a specific subproject or
source file but applying it globally seems safest.